feat added insert command
Added InsertCommand

Added services for better readability

TODO check for primary key when inserting

Added hash function for text

+++ src/helpers/headers/BinaryWriter.h
@@ -5,12 +5,14 @@
 #ifndef DBMS_BINARYWRITER_H
 #define DBMS_BINARYWRITER_H
 
+#include <fstream>
+#include <string>
 
 class BinaryWriter
 {
 public:
     void write_string(std::ofstream &file, const std::string& data) const;
-    void write_number(std::ofstream &file, const size_t data) const;
+    void write_number(std::ofstream &file, const long long data) const;
 };
